remove old bridge manager
